+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.9.9
 
+* added `start_at` to `IDAllocator`
+